refactor: make general sql mock trait
